Add discovery smap daemon
The discovery daemon is implemented as a smap driver. It currently detects new IP addresses on the network by tailing syslog, and looking for new DHCP leases -- therefore it must be run on the machine with the DHCP Server.

When it finds a new IP, it initates a scan of the device using nmap and the device detection scripts provided.  for each potential smap source that it finds, it creates a new config file in the config file repository, and commits that repo.

TODO: also hit Michael's service to spin up the daemon