Fix chunking issues when streaming through the window operator.
This could happen when you request a lot of data and pass it through
the window.  It happens since the we were not correctly passing the
current chunk region boundary to the windowing operator, for a couple
of reasons.  As a result, if the data were sampled asynchronously you
could get duplicate timestamps in the output if the chunk boundary was
unfortunate.  This fixes the root cause.