Make delete tags return value consistent
This patch changes the return value for deleting tags to include the
list of uuids which have been affected.  This is potentially somewhat
confusing, since you can't tell from the reply if the stream or the
tag has been deleted but I think it generally makes sense.

This also fixes the "has" operator to work on uuids -- formally,
"select * where has uuid" returned nothing, because the uuids are
stored separately from the metadata hstore.