Add the ability to set the default timezone
Currently, the only way to set the timezone is to set it explicitly on
every point, or go with America/Los_Angeles for everything.  This lets
you include a Properties/UnitofMeasure tag at the root of the sMAP
source, which sets the timezone which applies to all streams created
there.  For instance:

   [/]
   Properties/UnitofMeasure = America/New_York