Add noop option to queries.
This allows clients to attach a ?noop=1 flag to their requests to the
archiver.  The intention is that the queries return the same value as
the mutating query, but don't actually take the action.  This is
implemented by tweaking the compiled SQL submitted to postgers, so it
should be quite reliabile.  It's particularly handy when trying to
figure out what the effect of applying backreferences in a regex is.