add_actuator now creates a '_ts' timeseries for reporting
This timeseries will track the changes made to the actuator, providing
the actuator calls 'self.add(state)' within the get_state and set_state
methods. The timeseries will mirror the basic metadata of the actuator
stream (UnitOfTime, UnitOfMeasure, etc) and also have a pointer
Metadata/ActuatorPath that indicates the actuator path being tracked.
This will allow the basic reporting of actuator values while allowing
external clients to 'discover' the actuator path by reading the
metadata. This is most likely NOT a permanent solution.