Add additional table which allows for API keys
Currently, there are only "Subscription" keys.  These keys allow full
CRUD access to sMAP data associated with this key.  They can be used,
eg, on a sMAP source; however, they are also often used on clients in
order to access data.  However, there are several problems with them
that make them not great as api keys:

* only full CRUD access is possible
* there is no way to create multiple keys which can access the same
  data; this prevents any sort of user account scheme.
* there is also no way to restrict the permissions.

This commit adds an additional "permission" table; this allows for the
creation of new API keys (in the same format as current keys); these
keys may be restricted to select/apply, delete, or set permissions.
They can also be expired, and contain an M2M map to the underlying
subscriptions they grant access to.