Add ability to republish into a MongoDB collection
For a variety of reasons, you might want to mirror writes to a
MongoDB.  Although you could write a subscriber using the HTTP or
WebSockets API, this has the disavantage in that it's difficult
(basically impossible, I'd say) to guarantee you get all the incoming
data since if you ever disconnect, you can't reliabily tell what's
been received since you dropped off.

When you enable MongoDB, it adds each incoming object to the
smap.republisher collection in Mongo.  If the data can't be added, the
sMAP client add request will return an error, which should allow it to
buffer the data locally until the MongoDB connection is restored.

Because the incoming sMAP objects are somewhat difficult to use
directly, it's expected that there will be some background process for
renormalizing the data in MongoDB (or some other db) once it is
inserted.  However, that is beyond the scope of this patch, which
simply serves to basically tread MongoDB as a reliable FIFO for
integrating with other systems.