Add feature toggle for api keys
Since api keys require a db migration, we might not want to require
them (even though they are handy).  This requires the 'permissions'
feature to be set in the master configuration file before we generate
SQL that references the permissions tables.  As a result, updating to
this version should not break any existing installations.