Add postgres republish client
This client allows you add a new table to the archiver.  If the
republish option on the database is set to to true, all incoming data
will also be committed to this table in addition to being sent to
readingdb.  The incoming json is simply dumped in as text.

This is useful for sending the data to other endpoints.  sMAP doesn't
ever delete this table, so it may quickly become very large.  You
should ensure you are flushing it periodically.